Show action data after action creation
