+++ src/components/v5/common/ActionSidebar/ActionSidebar.tsx
@@ -4,7 +4,11 @@ import clsx from 'clsx';
 import Icon from '~shared/Icon';
 import { useMobile } from '~hooks';
 import { useActionSidebarContext } from '~context/ActionSidebarContext';
-import { useCloseSidebarClick, useGetActionData } from './hooks';
+import {
+  useCloseSidebarClick,
+  useGetActionData,
+  useRemoveTxParamOnClose,
+} from './hooks';
 import useToggle from '~hooks/useToggle';
 import Modal from '~v5/shared/Modal';
 import { ActionSidebarProps } from './types';
@@ -19,9 +23,8 @@ const ActionSidebar: FC<PropsWithChildren<ActionSidebarProps>> = ({
   children,
   transactionId,
 }) => {
-  const { defaultValues, loadingAction, isMotion } = useGetActionData(
-    transactionId || undefined,
-  );
+  const { defaultValues, loadingAction, isMotion } =
+    useGetActionData(transactionId);
 
   const {
     actionSidebarToggle: [
@@ -36,6 +39,7 @@ const ActionSidebar: FC<PropsWithChildren<ActionSidebarProps>> = ({
   const isMobile = useMobile();
 
   useDisableBodyScroll(isActionSidebarOpen);
+  useRemoveTxParamOnClose();
 
   return (
     <div
@@ -101,7 +105,7 @@ const ActionSidebar: FC<PropsWithChildren<ActionSidebarProps>> = ({
           transactionId={transactionId}
           formRef={formRef}
           defaultValues={defaultValues}
-          isMotion={isMotion || false}
+          isMotion={!!isMotion}
         />
       )}
