+++ src/redux/sagas/actions/editColony.ts
@@ -206,16 +206,10 @@ function* editColonyAction({
       meta,
     });
 
-    if (navigate) {
+    if (colonyName && navigate) {
       navigate(`/colony/${colonyName}/tx/${txHash}`, {
         state: { isRedirect: true },
       });
-    } else {
-      window.history.replaceState(
-        {},
-        '',
-        `${window.location.origin}${window.location.pathname}?tx=${txHash}`,
-      );
     }
   } catch (error) {
     return yield putError(ActionTypes.ACTION_EDIT_COLONY_ERROR, error, meta);
