+++ src/redux/sagas/actions/manageVerifiedRecipients.ts
@@ -246,18 +246,10 @@ function* manageVerifiedRecipients({
       meta,
     });
 
-    if (navigate) {
+    if (colonyName && navigate) {
       navigate(`/colony/${colonyName}/tx/${txHash}`, {
-        state: {
-          isRedirect: true,
-        },
+        state: { isRedirect: true },
       });
-    } else {
-      window.history.replaceState(
-        {},
-        '',
-        `${window.location.origin}${window.location.pathname}?tx=${txHash}`,
-      );
     }
   } catch (error) {
     return yield putError(
