+++ src/redux/sagas/actions/moveFunds.ts
@@ -135,12 +135,6 @@ function* createMoveFundsAction({
       navigate(`/colony/${colonyName}/tx/${txHash}`, {
         state: { isRedirect: true },
       });
-    } else {
-      window.history.replaceState(
-        {},
-        '',
-        `${window.location.origin}${window.location.pathname}?tx=${txHash}`,
-      );
     }
   } catch (caughtError) {
     putError(ActionTypes.ACTION_MOVE_FUNDS_ERROR, caughtError, meta);
