+++ src/redux/sagas/motions/moveFundsMotion.ts
@@ -214,18 +214,10 @@ function* moveFundsMotion({
       meta,
     });
 
-    if (colonyName) {
-      if (navigate) {
-        navigate(`/colony/${colonyName}/tx/${txHash}`, {
-          state: { isRedirect: true },
-        });
-      } else {
-        window.history.replaceState(
-          {},
-          '',
-          `${window.location.origin}${window.location.pathname}?tx=${txHash}`,
-        );
-      }
+    if (colonyName && navigate) {
+      navigate(`/colony/${colonyName}/tx/${txHash}`, {
+        state: { isRedirect: true },
+      });
     }
   } catch (caughtError) {
     putError(ActionTypes.MOTION_MOVE_FUNDS_ERROR, caughtError, meta);
